fix(widgets): rewrite Highlight as class-based Widget to remove circular dep
Convert Highlight.tsx (JSX function component) to Highlight.ts (class extending
Box), consistent with all other widgets in the package. This eliminates the need
for jsxImportSource in tsconfig and removes @termuijs/jsx from widgets dependencies,
which caused a circular build cycle (@termuijs/jsx already depends on @termuijs/widgets).
