+++ packages/widgets/package.json
@@ -32,7 +32,6 @@
   },
   "dependencies": {
     "@termuijs/core": "workspace:*",
-    "@termuijs/jsx": "workspace:*",
     "@termuijs/motion": "workspace:*"
   },
   "devDependencies": {
